+++ src/main/examples/BackdropPage.tsx
@@ -29,4 +29,4 @@ const BackdropPageBase = ({ settingValues, setSettingsControls }: IShowcaseBaseP
 	);
 }
 
-export const BackdropPage = withOptions<IBadgeProps>(BackdropPageBase, {}, 'BackdropPageBase');
+export const BackdropPage = withOptions<IBadgeProps>(BackdropPageBase, undefined, 'BackdropPageBase');